Restrict function return types to scalars only
  Reject array (memref) return types at three levels: Python signature
  validation raises TypeError at decoration time, C++ isValidReturnType
  rejects memref types, and the ctypes bridge error is simplified.

  Update 2D/3D array tests to return scalars instead of arrays, inlining
  the verify() wrapper pattern where needed.

+++ cpp/src/builders/MemRefBuilder.cpp
@@ -149,7 +149,7 @@ MemRefBuilder::flatToMultiIndex(int64_t flatIndex,
   int ndim = shape.size();
   llvm::SmallVector<int64_t, 4> indices(ndim);
   int64_t remaining = flatIndex;
-  for (int d = ndim - 1; d >= 0; --d) {
+  for (int64_t d = ndim - 1; d >= 0; --d) {
     indices[d] = remaining % shape[d];
     remaining /= shape[d];
   }
